+++ pywren_ibm_cloud/libs/openwhisk/client.py
@@ -26,30 +26,28 @@ from urllib.parse import urlparse
 logger = logging.getLogger(__name__)
 
 
-class CloudFunctionsClient:
+class OpenWhiskClient:
 
-    def __init__(self, endpoint, namespace, namespace_id=None, api_key=None,
-                 iam_api_key=None, token_manager=None, user_agent=None, insecure=False):
+    def __init__(self, endpoint, namespace, api_key=None, auth=None, insecure=False, user_agent=None):
         """
-        CloudFunctionsClient Constructor
+        OpenWhiskClient Constructor
+
+        :param endpoint: OpenWhisk endpoint.
+        :param namespace: User namespace.
+        :param api_key: User AUTH Key.  HTTP Basic authentication.
+        :param auth: Authorization token string "Basic eyJraWQiOiIyMDE5MDcyNCIsImFsZ...".
+        :param insecure: Insecure backend. Disable cert verification.
+        :param user_agent: User agent on requests.
         """
         self.endpoint = endpoint.replace('http:', 'https:')
         self.namespace = namespace
-        self.namespace_id = namespace_id
         self.api_key = api_key
-        self.iam_api_key = iam_api_key
-        self.token_manager = token_manager
+        self.auth = auth
 
         if self.api_key:
             api_key = str.encode(self.api_key)
             auth_token = base64.encodebytes(api_key).replace(b'\n', b'')
-            auth = 'Basic %s' % auth_token.decode('UTF-8')
-            self.effective_namespace = self.namespace
-
-        elif self.token_manager:
-            auth_token = token_manager._token
-            auth = 'Bearer ' + auth_token
-            self.effective_namespace = self.namespace_id
+            self.auth = 'Basic %s' % auth_token.decode('UTF-8')
 
         self.session = requests.session()
 
@@ -58,13 +56,14 @@ class CloudFunctionsClient:
             urllib3.disable_warnings(InsecureRequestWarning)
             self.session.verify = False
 
-        default_user_agent = self.session.headers['User-Agent']
-
         self.headers = {
             'content-type': 'application/json',
-            'Authorization': auth,
-            'User-Agent': default_user_agent + ' {}'.format(user_agent)
+            'Authorization': self.auth,
         }
+
+        if user_agent:
+            default_user_agent = self.session.headers['User-Agent']
+            self.headers['User-Agent'] = default_user_agent + ' {}'.format(user_agent)
 
         self.session.headers.update(self.headers)
         adapter = requests.adapters.HTTPAdapter()
@@ -90,7 +89,7 @@ class CloudFunctionsClient:
         data['exec'] = cfexec
 
         logger.debug('I am about to create a new cloud function action: {}'.format(action_name))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package,
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package,
                         action_name + "?overwrite=" + str(overwrite)])
 
         res = self.session.put(url, json=data)
@@ -107,7 +106,7 @@ class CloudFunctionsClient:
         Get an IBM Cloud Functions action
         """
         logger.debug("I am about to get a cloud function action: {}".format(action_name))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, action_name])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, action_name])
         res = self.session.get(url)
         return res.json()
 
@@ -116,7 +115,7 @@ class CloudFunctionsClient:
         List all IBM Cloud Functions actions in a package
         """
         logger.debug("I am about to list all actions from: {}".format(package))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, ''])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, ''])
         res = self.session.get(url)
         if res.status_code == 200:
             return res.json()
@@ -128,7 +127,7 @@ class CloudFunctionsClient:
         Delete an IBM Cloud Function
         """
         logger.debug("Delete cloud function action: {}".format(action_name))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, action_name])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, action_name])
         res = self.session.delete(url)
         resp_text = res.json()
 
@@ -137,7 +136,7 @@ class CloudFunctionsClient:
 
     def update_memory(self, package, action_name, memory):
         logger.debug('I am about to update the memory of the {} action to {}'.format(action_name, memory))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace,
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace,
                         'actions', package, action_name + "?overwrite=True"])
 
         data = {"limits": {"memory": memory}}
@@ -154,7 +153,7 @@ class CloudFunctionsClient:
         List all IBM Cloud Functions packages
         """
         logger.debug('I am about to list all the IBM CF packages')
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'packages'])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'packages'])
 
         res = self.session.get(url)
 
@@ -169,7 +168,7 @@ class CloudFunctionsClient:
         Delete an IBM Cloud Functions package
         """
         logger.debug("I am about to delete the package: {}".format(package))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'packages', package])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'packages', package])
         res = self.session.delete(url)
         resp_text = res.json()
 
@@ -183,7 +182,7 @@ class CloudFunctionsClient:
         Create an IBM Cloud Functions package
         """
         logger.debug('I am about to create the package {}'.format(package))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'packages', package + "?overwrite=False"])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'packages', package + "?overwrite=False"])
 
         data = {"name": package}
         res = self.session.put(url, json=data)
@@ -198,7 +197,7 @@ class CloudFunctionsClient:
         """
         Invoke an IBM Cloud Function by using new request.
         """
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, action_name])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, action_name])
         parsed_url = urlparse(url)
 
         try:
@@ -240,7 +239,7 @@ class CloudFunctionsClient:
         """
         Invoke an IBM Cloud Function waiting for the result.
         """
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions',
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions',
                         package, action_name + "?blocking=true&result=true"])
         resp = self.session.post(url, json=payload)
         result = resp.json()